Replaced double with BigDecimal

+++ Essentials/src/main/java/com/earth2me/essentials/ISettings.java
@@ -406,7 +406,7 @@ public interface ISettings extends IConf {
 
     boolean showZeroBaltop();
 
-    double getBaltopMinBalance();
+    BigDecimal getBaltopMinBalance();
 
     long getBaltopMinPlaytime();
